v0.0.2
Corrigidos diversos bugs, linhas faltando;
Alterado quase todos os tipos de variavel de RPGData, de double para int;

+++ VXAOS_Server/Code/RPGData/RPGActor.cs
@@ -1,13 +1,13 @@
 ﻿namespace VXAOS_Server.RPGData {
 	public class RPGActor:RPGBaseItem {
 		public string nickname = "";
-		public double class_id = 1;
-		public double initial_level = 1;
-		public double max_level = 99;
+		public int class_id = 1;
+		public int initial_level = 1;
+		public int max_level = 99;
 		public string character_name = "";
-		public double character_index = 0;
+		public int character_index = 0;
 		public string face_name = "";
-		public double face_index = 0;
-		public double[] equips = [0,0,0,0,0];
+		public int face_index = 0;
+		public int[] equips = [0,0,0,0,0];
 	}
 }
