+++ VXAOS_Server/Code/RPGData/RPGClassLearning.cs
@@ -1,7 +1,7 @@
 ﻿namespace VXAOS_Server.RPGData {
 	public class RPGClassLearning {
-		public double level = 1;
-		public double skill_id = 1;
+		public int level = 1;
+		public int skill_id = 1;
 		public string note = "";
 	}
 }
