+++ VXAOS_Server/Code/RPGData/RPGEnemyDropItem.cs
@@ -1,7 +1,7 @@
 ﻿namespace VXAOS_Server.RPGData {
 	public class RPGEnemyDropItem {
-		public double kind = 0;
-		public double data_id = 1;
-		public double denominator = 1;
+		public int kind = 0;
+		public int data_id = 1;
+		public int denominator = 1;
 	}
 }
