+++ VXAOS_Server/Code/RPGData/RPGEquipItem.cs
@@ -1,8 +1,8 @@
 ﻿namespace VXAOS_Server.RPGData {
 	public class RPGEquipItem:RPGBaseItem, IRPGItem {
-		public double price = 0;
-		public double etype_id = 0;
-		public double[] @params = [0,0,0,0,0,0,0,0];
+		public int price = 0;
+		public int etype_id = 0;
+		public int[] @params = [0,0,0,0,0,0,0,0];
 		public int level = 0;
 		public bool vip = false;
 		public bool soulbound = false;
@@ -10,7 +10,7 @@
 			return soulbound;
 		}
 		public int Price() {
-			return (int)price;
+			return price;
 		}
 	}
 	internal interface IRPGItem {
