+++ VXAOS_Server/Code/RPGData/RPGEventPageCondition.cs
@@ -6,13 +6,13 @@
 		public bool self_switch_valid = false;
 		public bool item_valid = false;
 		public bool actor_valid = false;
-		public double switch1_id = 1;
-		public double switch2_id = 1;
-		public double variable_id = 1;
-		public double variable_value = 0;
+		public int switch1_id = 1;
+		public int switch2_id = 1;
+		public int variable_id = 1;
+		public int variable_value = 0;
 		public string self_switch_ch = "A";
-		public double item_id = 1;
-		public double actor_id = 1;
+		public int item_id = 1;
+		public int actor_id = 1;
 
 	}
 }
