+++ VXAOS_Server/Code/RPGData/RPGEventPageGraphic.cs
@@ -1,10 +1,10 @@
 ﻿namespace VXAOS_Server.RPGData {
 	public class RPGEventPageGraphic {
-		public double tile_id = 0;
+		public int tile_id = 0;
 		public string character_name = "";
-		public double character_index = 0;
-		public double direction = 2;
-		public double pattern = 0;
+		public int character_index = 0;
+		public int direction = 2;
+		public int pattern = 0;
 
 	}
 }
