+++ VXAOS_Server/Code/RPGData/RPGMapEncounter.cs
@@ -1,11 +1,10 @@
 ﻿using Newtonsoft.Json;
-using System.Collections.Generic;
 
 namespace VXAOS_Server.RPGData {
 	public class RPGMapEncounter {
-		public double troop_id = 1;
-		public double weight = 10;
-		[JsonConverter(typeof(ListConverter<double>))]
-		public List<double> region_set = new List<double>();
+		public int troop_id = 1;
+		public int weight = 10;
+		[JsonConverter(typeof(ListConverter<int>))]
+		public List<int> region_set = new List<int>();
 	}
 }
