+++ VXAOS_Server/Code/RPGData/RPGMapInfo.cs
@@ -1,11 +1,11 @@
 ﻿namespace VXAOS_Server.RPGData {
 	public class RPGMapInfo {
 		public string name = "";
-		public double parent_id = 0;
-		public double order = 0;
+		public int parent_id = 0;
+		public int order = 0;
 		public bool expanded = false;
-		public double scroll_x = 0;
-		public double scroll_y = 0;
+		public int scroll_x = 0;
+		public int scroll_y = 0;
 
 	}
 }
