+++ VXAOS_Server/Code/RPGData/RPGSystemTestBattler.cs
@@ -1,7 +1,7 @@
 ﻿namespace VXAOS_Server.RPGData {
 	public class RPGSystemTestBattler {
-		public double actor_id = 1;
-		public double level = 1;
-		public double[] equips = [0,0,0,0,0];
+		public int actor_id = 1;
+		public int level = 1;
+		public int[] equips = [0,0,0,0,0];
 	}
 }
