+++ VXAOS_Server/Code/RPGData/RPGTroopMember.cs
@@ -1,8 +1,8 @@
 ﻿namespace VXAOS_Server.RPGData {
 	public class RPGTroopMember {
-		public double enemy_id = 1;
-		public double x = 0;
-		public double y = 0;
+		public int enemy_id = 1;
+		public int x = 0;
+		public int y = 0;
 		public bool hidden = false;
 
 	}
