+++ VXAOS_Server/Code/RPGData/RPGTroopPageCondition.cs
@@ -5,13 +5,13 @@
 		public bool enemy_valid = false;
 		public bool actor_valid = false;
 		public bool switch_valid = false;
-		public double turn_a = 0;
-		public double turn_b = 0;
-		public double enemy_index = 0;
-		public double enemy_hp = 50;
-		public double actor_id = 1;
-		public double actor_hp = 50;
-		public double switch_id = 1;
+		public int turn_a = 0;
+		public int turn_b = 0;
+		public int enemy_index = 0;
+		public int enemy_hp = 50;
+		public int actor_id = 1;
+		public int actor_hp = 50;
+		public int switch_id = 1;
 
 	}
 }
