+++ VXAOS_Server/Code/RPGData/RPGUsableItemDamage.cs
@@ -5,10 +5,10 @@ using System.Text.RegularExpressions;
 
 namespace VXAOS_Server.RPGData {
 	public class RPGUsableItemDamage {
-		public double type = 0;
-		public double element_id = 0;
+		public int type = 0;
+		public int element_id = 0;
 		public string formula = "0";
-		public double variance = 20;
+		public int variance = 20;
 		public bool critical = false;
       public bool IsNone() {
          return type == 0;
